Move lastParsedMessage logic into PeerMessageSender
